+++ Krowi_AchievementFilter.lua
@@ -34,7 +34,7 @@ loadHelper:RegisterEvent("ACHIEVEMENT_EARNED");
 local function LoadKrowi_AchievementFilter()
     addon.Diagnostics.Load();
 
-    addon.Data.ExportedCategories.InjectDynamicOptions();
+    addon.Data.SpecialCategories.InjectDynamicOptions();
     addon.Data.ExportedCalendarEvents.InjectDynamicOptions();
     if addon.Data.ExportedWidgetEvents then
         addon.Data.ExportedWidgetEvents.InjectDynamicOptions();